# openapi.yaml
openapi: 3.0.3
info:
  title: Mon École API
  version: 1.0.0
  description: School management system API.
paths:
...
  /api/v1/enrollments/{id}/change-class/:
    post:
      operationId: enrollments_change_class_create
      description: |-
        CRUD for school-owned records: scoped to the current school, audited on every write.

        Subclasses set `queryset`, `serializer_class`, `required_permissions` and `audit_module`.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this enrolment.
        required: true
      tags:
      - enrollments
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeClassRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ChangeClassRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ChangeClassRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enrollment'
          description: ''
...
  /api/v1/guardians/:
    get:
      operationId: guardians_list
      description: |-
        CRUD for school-owned records: scoped to the current school, audited on every write.

        Subclasses set `queryset`, `serializer_class`, `required_permissions` and `audit_module`.
      parameters:
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      tags:
      - guardians
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedGuardianList'
          description: ''
...
components:
  schemas:
...
    ChangeClassRequest:
      type: object
      properties:
        class_group:
          type: integer
        date:
          type: string
          format: date
        reason:
          type: string
          default: ''
          maxLength: 255
      required:
      - class_group
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        student:
          allOf:
          - $ref: '#/components/schemas/EnrollmentStudent'
          readOnly: true
        academic_year:
          type: integer
          readOnly: true
        academic_year_name:
          type: string
          readOnly: true
        class_group:
          type: integer
          readOnly: true
        class_name:
          type: string
          readOnly: true
        level_name:
          type: string
          readOnly: true
        enrollment_date:
          type: string
          format: date
        kind:
          $ref: '#/components/schemas/KindEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/EnrollmentStatusEnum'
          readOnly: true
        previous_school:
          type: string
          maxLength: 200
        ended_on:
          type: string
          format: date
          readOnly: true
          nullable: true
        end_reason:
          type: string
          readOnly: true
        transfer_to:
          type: string
          readOnly: true
          description: School the student left for.
        notes:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - academic_year
      - academic_year_name
      - class_group
      - class_name
      - created_at
      - end_reason
      - ended_on
      - enrollment_date
      - id
      - level_name
      - status
      - student
      - transfer_to
...
    EnrollmentStatusEnum:
      enum:
      - active
      - class_changed
      - withdrawn
      - completed
      - cancelled
      type: string
      description: |-
        * `active` - Active
        * `class_changed` - Moved to another class
        * `withdrawn` - Left the school
        * `completed` - Year completed
        * `cancelled` - Cancelled
    EnrollmentStudent:
      type: object
      properties:
        id:
          type: integer
        student_number:
          type: string
        full_name:
          type: string
        gender:
          type: string
        date_of_birth:
          type: string
          format: date
          nullable: true
        photo_url:
          type: string
          nullable: true
      required:
      - date_of_birth
      - full_name
      - gender
      - id
      - photo_url
      - student_number
...
    Guardian:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        first_name:
          type: string
          maxLength: 100
        last_name:
          type: string
          maxLength: 100
        full_name:
          type: string
          readOnly: true
        phone:
          type: string
          maxLength: 30
        alt_phone:
          type: string
          maxLength: 30
        email:
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        address:
          type: string
        occupation:
          type: string
          maxLength: 100
        students:
          type: array
          items:
            $ref: '#/components/schemas/GuardianStudent'
          readOnly: true
      required:
      - first_name
      - full_name
      - id
      - last_name
      - students
...
    GuardianStudent:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
        relationship:
          type: string
      required:
      - full_name
      - id
      - relationship
...
    KindEnum:
      enum:
      - new
      - re_enrolment
      - transfer_in
      type: string
      description: |-
        * `new` - New student
        * `re_enrolment` - Re-enrolment
        * `transfer_in` - Transfer from another school
...
    PaginatedGuardianList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Guardian'
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT